+++ src/modules/tabs/swiper/SwiperScreen.js
@@ -14,7 +14,7 @@ import {
 } from 'react-native';
 
 // first party
-import { COLORS } from '../../../common/helpers/styles';
+import * as STYLES from '../../../constants/Styles';
 
 // ========================================================
 // Component
@@ -47,7 +47,7 @@ const styles = StyleSheet.create({
         flex: 1,
         justifyContent: 'center',
         alignItems: 'center',
-        backgroundColor: COLORS.TUMBLEWEED_BROWN,
+        backgroundColor: STYLES.COLORS.TUMBLEWEED_BROWN,
     },
 });